Replace OpenSSL with PolarSSL for use with libcurl. See issue #3418
The OpenSSL licence is not compatible with GPL3 unless an exception is added to the GPL3 licence, which requires permission from the copyright owners. PolarSSL is GPL2 (or later), so the licences are compatible. Curl 7.36.0 and PolarSSL 1.3.7 static libraries have been included to provide SSL support for tdm_update and in-game downloads.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@6012 49c82d7f-2e2a-0410-a16f-ae8f201b507f